Refactor: 네이밍 규칙에 맞게 controller 이름 변경
